+++ lib/database_execution.rb
@@ -1,0 +1,5 @@
+module DatabaseExecution
+  def execute_sql(sql)
+    ActiveRecord::Base.connection.execute(sql)
+  end
+end
